tests/wpt/tests/html/canvas/element/drawing-images-to-the-canvas/2d.drawImage.detachedcanvas.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.drawImage.detachedcanvas</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.drawImage.detachedcanvas</h1>
<p class="desc">drawImage with detached OffscreenCanvas as the source should throw exception</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("drawImage with detached OffscreenCanvas as the source should throw exception");
_addTest(function(canvas, ctx) {

  var canvas2 = new OffscreenCanvas(80, 80);
  (new MessageChannel()).port1.postMessage(canvas2, [canvas2]);
  assert_throws_dom("INVALID_STATE_ERR", function() { ctx.drawImage(canvas2, 0, 0); });

});
</script>
